Move Cas3BookingSearchService to CAS3 namespace
